Remove vite helper
