+++ tests/View/Blade/BladeHelpersTest.php
@@ -11,6 +11,6 @@ class BladeHelpersTest extends AbstractBladeTestCase
         $this->assertSame('<?php dd($var1); ?>', $this->compiler->compileString('@dd($var1)'));
         $this->assertSame('<?php dd($var1, $var2); ?>', $this->compiler->compileString('@dd($var1, $var2)'));
         $this->assertSame('<?php dump($var1, $var2); ?>', $this->compiler->compileString('@dump($var1, $var2)'));
-        $this->assertSame('<?php echo vite([\'resources/js/app.js\']); ?>', $this->compiler->compileString('@vite([\'resources/js/app.js\'])'));
+        $this->assertSame('<?php echo app(\'Illuminate\Foundation\Vite\')([\'resources/js/app.js\']); ?>', $this->compiler->compileString('@vite([\'resources/js/app.js\'])'));
     }
 }
